% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_dist.R
\name{bm_dist}
\alias{bm_dist}
\title{Exakte verteilung von Brunner-Munzel's W}
\usage{
bm_dist(data)
}
\arguments{
\item{data}{Output von `npvTools::exact_bm_dist`}
}
\description{
Diese Funktion nimmt den Output von `npvTools::exact_bm_dist`
als Input und erstellt einen Plot der exakten Verteilung
}
\examples{
\dontrun{
  test_data <- data.frame(
    value = c(0, 0, 1, 1, 1, 1, 1, 1, 3, 3),
    group = factor(rep(1:2, each = 5))
  )
  real_data <- npvTools::exact_bm_dist(test_data)
  bm_dist(real_data)
}

}
