% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_plot.R
\name{kendall.plot}
\alias{kendall.plot}
\title{Plot der exakten Verteilung von S (Kendall-Test)}
\usage{
kendall.plot(data)
}
\arguments{
\item{data}{Ein Dataframe oder eine Matrix mit mindestens zwei Spalten,
welches im Skript verwendet wird, um die Häufigkeitstabelle zu erstellen.
Die erste Spalte muss die Werte der Statistik S enthalten,
die zweite Spalte die zugehörigen Häufigkeiten.}
}
\value{
Ein ggplot2-Objekt, das das Balkendiagramm darstellt.
}
\description{
Diese Funktion erstellt ein Balkendiagramm, das die Häufigkeitsverteilung
der Teststatistik S für den Kendall-Test anzeigt.
}
