% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_plot.R
\name{norm.plot}
\alias{norm.plot}
\title{z-Verteilung Plot}
\usage{
norm.plot(
  emp,
  mean,
  var,
  tail = c("less", "greater", "two.sided"),
  xlab = (xmax/10),
  ylab = 0.001
)
}
\arguments{
\item{emp}{Empirisches z}

\item{mean}{Mittelwert der Verteilung}

\item{var}{Varianz der Verteilung}

\item{tail}{Indiziert die Richtung der Hypothese}

\item{xlab}{Move the annotation on the x axis}

\item{ylab}{Move the annotation on the y axis}

\item{xmax}{Definiert Breite des Plots. 4 sieht meistens gut aus.}
}
\description{
Plottet die z Verteilung
}
