% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tbl.R
\name{read_tbl}
\alias{read_tbl}
\title{Tabelle aus Excel einlesen (für Typst-Output wird die _typst-Version verwendet)}
\usage{
read_tbl(filename, html = knitr::is_html_output())
}
\arguments{
\item{filename}{Pfad relativ zum Projekt-Root, z.B. "Tables/binom1.xlsx"}

\item{html}{Wird HTML gerendert? Standard: knitr::is_html_output()}
}
\description{
Tabelle aus Excel einlesen (für Typst-Output wird die _typst-Version verwendet)
}
