% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_dist.R
\name{wp_dist}
\alias{wp_dist}
\title{Exakte Verteilung von W-plus}
\usage{
wp_dist(data)
}
\arguments{
\item{data}{dataframe mit der exakten Verteilung von W-plus}
}
\description{
Plottet die exakte Verteilung von W-plus
}
