% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl2hux.R
\name{xl2hux}
\alias{xl2hux}
\title{Excel-Datei einlesen und als formatierte Huxtable-Tabelle ausgeben}
\usage{
xl2hux(filename)
}
\arguments{
\item{filename}{Zeichenkette. Name der Excel-Datei, die sich im Ordner "Tables"
relativ zum Projektwurzelverzeichnis befindet.}
}
\value{
Ein huxtable-Objekt mit dem eingelesenen und formatierten Tabelleninhalt.
}
\description{
Diese Funktion liest eine Excel-Datei aus dem Unterordner "Tables" des Projektordners
ein. Die Datei wird ohne Spaltennamen importiert und in ein huxtable-Objekt umgewandelt
und grob formatiert. Es werden keine Spalten- oder Zeilennamen angezeigt, die
Spaltenbreite wird festgelegt, die Schriftart auf "Times New Roman" gesetzt, die
Schriftgrösse auf 14 und der Text zentriert.
}
