% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot.R
\name{z.plot}
\alias{z.plot}
\title{z-Verteilung Plot}
\usage{
z.plot(
  z,
  tail = c("less", "greater", "two.sided"),
  xmax = 4,
  xlab = (xmax/10),
  ylab = 0.001
)
}
\arguments{
\item{z}{Empirisches z}

\item{tail}{Indiziert die Richtung der Hypothese}

\item{xmax}{Definiert Breite des Plots. 4 sieht meistens gut aus.}

\item{xlab}{Move the annotation on the x axis}

\item{ylab}{Move the annotation on the y axis}
}
\description{
Plottet die z Verteilung
}
